moved matrix-related operations from cxoperations.hpp to separate cxmat.hpp. removed obsolete cvmat.hpp. CvMat::step is not set to 0 for single-row matrices anymore. Fixed bug in gemm when one of input matrices could be suddenly deallocated.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@1821 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cxcore/cxmatrix.cpp
@@ -520,11 +520,11 @@ void reduce(const Mat& src, Mat& dst, int dim, int op, int dtype)
             if(sdepth == CV_16U && ddepth == CV_32F)
                 func = reduceR_<ushort,float,OpAdd<float> >;
             if(sdepth == CV_16U && ddepth == CV_64F)
-                func = reduceR_<ushort,double,OpAdd<float> >;
+                func = reduceR_<ushort,double,OpAdd<double> >;
             if(sdepth == CV_16S && ddepth == CV_32F)
                 func = reduceR_<short,float,OpAdd<float> >;
             if(sdepth == CV_16S && ddepth == CV_64F)
-                func = reduceR_<short,double,OpAdd<float> >;
+                func = reduceR_<short,double,OpAdd<double> >;
             if(sdepth == CV_32F && ddepth == CV_32F)
                 func = reduceR_<float,float,OpAdd<float> >;
             if(sdepth == CV_32F && ddepth == CV_64F)
@@ -564,11 +564,11 @@ void reduce(const Mat& src, Mat& dst, int dim, int op, int dtype)
             if(sdepth == CV_16U && ddepth == CV_32F)
                 func = reduceC_<ushort,float,OpAdd<float> >;
             if(sdepth == CV_16U && ddepth == CV_64F)
-                func = reduceC_<ushort,double,OpAdd<float> >;
+                func = reduceC_<ushort,double,OpAdd<double> >;
             if(sdepth == CV_16S && ddepth == CV_32F)
                 func = reduceC_<short,float,OpAdd<float> >;
             if(sdepth == CV_16S && ddepth == CV_64F)
-                func = reduceC_<short,double,OpAdd<float> >;
+                func = reduceC_<short,double,OpAdd<double> >;
             if(sdepth == CV_32F && ddepth == CV_32F)
                 func = reduceC_<float,float,OpAdd<float> >;
             if(sdepth == CV_32F && ddepth == CV_64F)
